split out helper methods into appropriate services
